Corregir tipo de tema_id en migraciones: unsignedInteger para coincidir con temas.id

+++ database/migrations/2026_02_13_000001_create_subtemas_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('subtemas', function (Blueprint $table) {
             $table->id();
             $table->string('nombre');
-            $table->unsignedBigInteger('tema_id');
+            $table->unsignedInteger('tema_id');
             $table->foreign('tema_id')->references('id')->on('temas')->onDelete('cascade');
         });
     }
